// src/app/module/recruiter/recruiter.interface.ts
export interface IUpdateRecruiterPayload {
    recruiter?: {
        name?: string;
        profilePhoto?: string;
        contactNumber?: string;
        companyName?: string;
        companyLogo?: string;
        companyWebsite?: string;
        companyAddress?: string;
        designation?: string;
        industry?: string;
        companySize?: string;
        description?: string;
    }
}